move user related components to new lib layout
