+++ projects/storefrontlib/src/cms-components/myaccount/forgot-password/forgot-password.component.html
