+++ projects/storefrontlib/src/cms-components/myaccount/index.ts
@@ -1,8 +1,12 @@
 export * from './address-book/index';
 export * from './close-account/index';
+export * from './forgot-password/forgot-password.component';
+export * from './forgot-password/forgot-password.module';
 export * from './order/index';
 export * from './payment-methods/payment-methods.component';
 export * from './payment-methods/payment-methods.module';
+export * from './reset-password/reset-password-form.component';
+export * from './reset-password/reset-password.module';
 export * from './update-email/index';
 export * from './update-password/index';
 export * from './update-profile/index';
