+++ projects/storefrontlib/src/cms-components/myaccount/reset-password/reset-password-form.component.ts
@@ -7,7 +7,7 @@ import {
 } from '@angular/forms';
 import { RoutingService, UserService } from '@spartacus/core';
 import { Subscription } from 'rxjs';
-import { CustomFormValidators } from '../../ui/validators/custom-form-validators';
+import { CustomFormValidators } from '../../../lib/ui/validators/custom-form-validators';
 
 @Component({
   selector: 'cx-reset-password-form',
