+++ projects/storefrontlib/src/cms-components/user/index.ts
@@ -1,2 +1,9 @@
+export * from './login-form/login-form.component';
+export * from './login-form/login-form.module';
+export * from './login/login.component';
+export * from './login/login.module';
 export * from './logout/logout-guard';
 export * from './logout/logout.module';
+export * from './register/register.component';
+export * from './register/register.module';
+export * from './user.module';
