+++ projects/storefrontlib/src/cms-components/user/register/register.component.spec.ts
