--- projects/storefrontlib/src/lib/user/index.ts
@@ -1,9 +1,0 @@
-export * from './user.module';
-
-export * from './login/login.module';
-export * from './login/login.component';
-
-export * from './login-form/login-form.module';
-export * from './login-form/login-form.component';
-
-export * from './register/register.component';
